+++ tests/spec/directives/forward/member/import/import_to_forward/transitive.rs
@@ -42,7 +42,6 @@ mod transitive {
     }
 
     #[test]
-    #[ignore] // wrong result
     fn function() {
         let runner = runner().with_cwd("function");
         assert_eq!(
@@ -54,7 +53,6 @@ mod transitive {
         );
     }
     #[test]
-    #[ignore] // unexepected error
     fn mixin() {
         let runner = runner().with_cwd("mixin");
         assert_eq!(
@@ -66,7 +64,6 @@ mod transitive {
         );
     }
     #[test]
-    #[ignore] // unexepected error
     fn variable() {
         let runner = runner().with_cwd("variable");
         assert_eq!(