+++ tests/spec/directives/import/mod.rs
@@ -11,6 +11,8 @@ mod error;
 
 mod escaped;
 
+mod implicit_dependencies;
+
 mod load;
 
 mod nested;