+++ tests/spec/main.rs
@@ -1,5 +1,5 @@
 //! Tests auto-converted from "sass-spec/spec"
-//! version db56987c, 2021-06-02 14:22:39 -0700.
+//! version f595d351, 2021-06-14 14:20:03 -0700.
 //! See <https://github.com/sass/sass-spec> for source material.\n
 //! The following tests are excluded from conversion:
 //! ["core_functions/selector/extend", "core_functions/selector/is_superselector", "core_functions/selector/unify", "directives/extend", "libsass-closed-issues/issue_1801", "libsass-todo-issues/issue_1801", "libsass-todo-issues/issue_221262.hrx", "libsass-todo-issues/issue_221260.hrx", "libsass-todo-issues/issue_221292.hrx", "libsass/unicode-bom/utf-16-big", "libsass/unicode-bom/utf-16-little", "non_conformant/scss/huge.hrx", "non_conformant/scss/multiline_var.hrx"]
@@ -25,6 +25,8 @@ mod libsass_todo_issues;
 mod libsass_todo_tests;
 
 mod non_conformant;
+
+mod operators;
 
 mod parser;